// src/types/ProcessData.ts
// src/types/ProcessData.ts
export interface ProcessData {
    stack: string;
    methodology: string;
    frequencyToDeploy: string;
    latamInfluence: string;
    accountabilityLevel: string;
    observations: string;
    status: string; // Por ejemplo: "Good", "Warning", "Bad", "Not Defined"
  }
  